backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2
drf-yasg==1.21.8
psycopg2-binary==2.9.9
redis==5.2.1
gunicorn==23.0.0
firebase-admin==6.5.0
